requirements.txt
streamlit==1.40.0
pandas==2.2.0
numpy==1.26.0
matplotlib==3.8.0
Pillow==10.0.0
fpdf==1.7.2
mplsoccer==1.1.12
scipy==1.11.3
soccerdata==0.5.2